moved classes and implementations into h/cpp files. no warnings. added changes to .replit
